+++ docs/en_US/changelog.md
@@ -1,4 +1,12 @@
 #Changelog : Netatmo OpenData
 
+## 24 may 2020
+- Adjust widget size on creation
+- Widget's optimization for V3 
+- Change update time value, with Netatmo value UTC
+
+## 23 may 2020
+- Add new data  : Gust Strength,  Gust Angle, Rain on the last hour and rain for the day
+
 ## 16 may 2020
 - First release in 'beta'