+++ docs/fr_FR/changelog.md
@@ -1,5 +1,13 @@
 #Changelog : Netatmo OpenData
 
+## 24 mai 2020 
+- Ajustement de la taille des widgets à la création
+- Optimisation des widgets pour V3
+- Remplacement de la date de récupération des valeurs, par celle fournis par Netatmo UTC
+
+## 23 mai 2020
+- Ajout de nouvelles données : rafales de vent (vitesse et direction), pluie sur 1h, pluie sur la journée
+
 ## 16 mai 2020
 - 1ère version in 'beta'